:card_file_box: Update all pricing column type to integer

+++ database/migrations/2020_00_03_000003_create_order_items_table.php
@@ -23,7 +23,7 @@ class CreateOrderItemsTable extends Migration
             $table->string('sku')->nullable()->index();
             $table->morphs('product');
             $table->integer('quantity');
-            $table->decimal('unit_price_amount', 10, 2);
+            $table->integer('unit_price_amount');
 
             $this->addForeignKey($table, 'order_id', $this->getTableName('orders'));
         });
